Moved the hints to Translations.cs for easier usage
